amelie/requirements.txt
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
tqdm==4.66.1
openai-whisper==20231117